backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-multipart
pymupdf
openai
chromadb
weaviate-client
telethon
